+++ core/tauri/src/lib.rs
@@ -65,8 +65,7 @@ pub use cocoa;
 #[cfg(target_os = "macos")]
 #[doc(hidden)]
 pub use embed_plist;
-/// The Tauri error enum.
-pub use error::Error;
+pub use error::{Error, Result};
 #[cfg(target_os = "ios")]
 #[doc(hidden)]
 pub use swift_rs;
@@ -74,10 +73,7 @@ pub use swift_rs;
 pub use tauri_macros::mobile_entry_point;
 pub use tauri_macros::{command, generate_handler};
 
-pub mod api;
 pub(crate) mod app;
-#[cfg(feature = "protocol-asset")]
-pub(crate) mod asset_protocol;
 pub mod async_runtime;
 pub mod command;
 mod error;
@@ -86,6 +82,7 @@ pub mod ipc;
 mod manager;
 mod pattern;
 pub mod plugin;
+pub(crate) mod protocol;
 mod vibrancy;
 pub mod window;
 use tauri_runtime as runtime;
@@ -106,6 +103,8 @@ mod state;
 #[cfg_attr(doc_cfg, doc(cfg(all(desktop, feature = "tray-icon"))))]
 pub mod tray;
 pub use tauri_utils as utils;
+
+pub use http;
 
 /// A Tauri [`Runtime`] wrapper around wry.
 #[cfg(feature = "wry")]
@@ -161,9 +160,6 @@ pub use plugin::mobile::{handle_android_plugin_response, send_channel_data};
 #[doc(hidden)]
 pub use tauri_runtime_wry::wry;
 
-/// `Result<T, ::tauri::Error>`
-pub type Result<T> = std::result::Result<T, Error>;
-
 /// A task to run on the main thread.
 pub type SyncTask = Box<dyn FnOnce() + Send>;
 
@@ -173,9 +169,6 @@ use std::{
   fmt::{self, Debug},
   sync::Arc,
 };
-
-// Export types likely to be used by the application.
-pub use runtime::http;
 
 #[cfg(feature = "wry")]
 #[cfg_attr(doc_cfg, doc(cfg(feature = "wry")))]
@@ -188,7 +181,7 @@ pub use runtime::ActivationPolicy;
 #[cfg(target_os = "macos")]
 pub use self::utils::TitleBarStyle;
 
-pub use self::event::{Event, EventHandler};
+pub use self::event::{Event, EventId};
 pub use {
   self::app::{
     App, AppHandle, AssetResolver, Builder, CloseRequestApi, GlobalWindowEvent, RunEvent,
@@ -242,7 +235,7 @@ pub fn log_stdout() {
       if let Ok(len) = reader.read_line(&mut buffer) {
         if len == 0 {
           break;
-        } else if let Ok(msg) = CString::new(buffer.clone())
+        } else if let Ok(msg) = CString::new(buffer.as_bytes())
           .map_err(|_| ())
           .and_then(|c| c.into_string().map_err(|_| ()))
         {
@@ -616,7 +609,7 @@ pub trait Manager<R: Runtime>: sealed::ManagerBase<R> {
   ///   })
   ///   .invoke_handler(tauri::generate_handler![synchronize]);
   /// ```
-  fn listen_global<F>(&self, event: impl Into<String>, handler: F) -> EventHandler
+  fn listen_global<F>(&self, event: impl Into<String>, handler: F) -> EventId
   where
     F: Fn(Event) + Send + 'static,
   {
@@ -626,7 +619,7 @@ pub trait Manager<R: Runtime>: sealed::ManagerBase<R> {
   /// Listen to a global event only once.
   ///
   /// See [`Self::listen_global`] for more information.
-  fn once_global<F>(&self, event: impl Into<String>, handler: F) -> EventHandler
+  fn once_global<F>(&self, event: impl Into<String>, handler: F)
   where
     F: FnOnce(Event) + Send + 'static,
   {
@@ -681,8 +674,8 @@ pub trait Manager<R: Runtime>: sealed::ManagerBase<R> {
   ///     Ok(())
   ///   });
   /// ```
-  fn unlisten(&self, handler_id: EventHandler) {
-    self.manager().unlisten(handler_id)
+  fn unlisten(&self, id: EventId) {
+    self.manager().unlisten(id)
   }
 
   /// Fetch a single window from the manager.
@@ -824,13 +817,13 @@ pub trait Manager<R: Runtime>: sealed::ManagerBase<R> {
   }
 
   /// Gets the scope for the IPC.
-  fn ipc_scope(&self) -> IpcScope {
+  fn ipc_scope(&self) -> scope::ipc::Scope {
     self.state::<Scopes>().inner().ipc.clone()
   }
 
   /// Gets the scope for the asset protocol.
   #[cfg(feature = "protocol-asset")]
-  fn asset_protocol_scope(&self) -> FsScope {
+  fn asset_protocol_scope(&self) -> scope::fs::Scope {
     self.state::<Scopes>().inner().asset_protocol.clone()
   }
 
@@ -919,7 +912,8 @@ macro_rules! run_main_thread {
     let (tx, rx) = channel();
     let self_ = $self.clone();
     let task = move || {
-      let _ = tx.send($ex(self_));
+      let f = $ex;
+      let _ = tx.send(f(self_));
     };
     $self.app_handle.run_on_main_thread(Box::new(task))?;
     rx.recv().map_err(|_| crate::Error::FailedToReceiveMessage)