+++ core/tauri/src/tray.rs
@@ -11,15 +11,105 @@ use crate::menu::ContextMenu;
 use crate::menu::MenuEvent;
 use crate::{run_main_thread, AppHandle, Icon, Manager, Runtime};
 use std::path::Path;
-pub use tray_icon::{ClickType, Rectangle, TrayIconEvent, TrayIconId};
+pub use tray_icon::TrayIconId;
 
 // TODO(muda-migration): figure out js events
+
+/// Describes a rectangle including position (x - y axis) and size.
+#[derive(Debug, PartialEq, Clone, Copy, Default)]
+pub struct Rectangle {
+  /// The x-coordinate of the upper-left corner of the rectangle.
+  pub left: f64,
+  /// The y-coordinate of the upper-left corner of the rectangle.
+  pub top: f64,
+  /// The x-coordinate of the lower-right corner of the rectangle.
+  pub right: f64,
+  /// The y-coordinate of the lower-right corner of the rectangle.
+  pub bottom: f64,
+}
+
+/// Describes the click type that triggered this tray icon event.
+#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+pub enum ClickType {
+  /// Left mouse click.
+  Left,
+  /// Right mouse click.
+  Right,
+  /// Double left mouse click.
+  Double,
+}
+
+impl Default for ClickType {
+  fn default() -> Self {
+    Self::Left
+  }
+}
+
+/// Describes a tray event emitted when a tray icon is clicked
+///
+/// ## Platform-specific:
+///
+/// - **Linux**: Unsupported. The event is not emmited even though the icon is shown,
+/// the icon will still show a context menu on right click.
+#[derive(Debug, Clone, Default)]
+pub struct TrayIconEvent {
+  /// Id of the tray icon which triggered this event.
+  pub id: TrayIconId,
+  /// Physical X Position of the click the triggered this event.
+  pub x: f64,
+  /// Physical Y Position of the click the triggered this event.
+  pub y: f64,
+  /// Position and size of the tray icon
+  pub icon_rect: Rectangle,
+  /// The click type that triggered this event.
+  pub click_type: ClickType,
+}
+
+impl TrayIconEvent {
+  /// Returns the id of the tray icon which triggered this event.
+  pub fn id(&self) -> &TrayIconId {
+    &self.id
+  }
+}
+
+impl From<tray_icon::Rectangle> for Rectangle {
+  fn from(value: tray_icon::Rectangle) -> Self {
+    Self {
+      bottom: value.bottom,
+      left: value.left,
+      top: value.top,
+      right: value.right,
+    }
+  }
+}
+
+impl From<tray_icon::ClickType> for ClickType {
+  fn from(value: tray_icon::ClickType) -> Self {
+    match value {
+      tray_icon::ClickType::Left => Self::Left,
+      tray_icon::ClickType::Right => Self::Right,
+      tray_icon::ClickType::Double => Self::Double,
+    }
+  }
+}
+
+impl From<tray_icon::TrayIconEvent> for TrayIconEvent {
+  fn from(value: tray_icon::TrayIconEvent) -> Self {
+    Self {
+      id: value.id,
+      x: value.x,
+      y: value.y,
+      icon_rect: value.icon_rect.into(),
+      click_type: value.click_type.into(),
+    }
+  }
+}
 
 /// [`TrayIcon`] builder struct and associated methods.
 #[derive(Default)]
 pub struct TrayIconBuilder<R: Runtime> {
   on_menu_event: Option<GlobalMenuEventListener<AppHandle<R>>>,
-  on_tray_event: Option<GlobalTrayIconEventListener<TrayIcon<R>>>,
+  on_tray_icon_event: Option<GlobalTrayIconEventListener<TrayIcon<R>>>,
   inner: tray_icon::TrayIconBuilder,
 }
 
@@ -34,7 +124,7 @@ impl<R: Runtime> TrayIconBuilder<R> {
     Self {
       inner: tray_icon::TrayIconBuilder::new(),
       on_menu_event: None,
-      on_tray_event: None,
+      on_tray_icon_event: None,
     }
   }
 
@@ -133,11 +223,11 @@ impl<R: Runtime> TrayIconBuilder<R> {
   }
 
   /// Set a handler for this tray icon events.
-  pub fn on_tray_event<F: Fn(&TrayIcon<R>, TrayIconEvent) + Sync + Send + 'static>(
+  pub fn on_tray_icon_event<F: Fn(&TrayIcon<R>, TrayIconEvent) + Sync + Send + 'static>(
     mut self,
     f: F,
   ) -> Self {
-    self.on_tray_event.replace(Box::new(f));
+    self.on_tray_icon_event.replace(Box::new(f));
     self
   }
 
@@ -157,7 +247,11 @@ impl<R: Runtime> TrayIconBuilder<R> {
       app_handle: manager.app_handle().clone(),
     };
 
-    icon.register(&icon.app_handle, self.on_menu_event, self.on_tray_event);
+    icon.register(
+      &icon.app_handle,
+      self.on_menu_event,
+      self.on_tray_icon_event,
+    );
 
     Ok(icon)
   }
@@ -195,7 +289,7 @@ impl<R: Runtime> TrayIcon<R> {
     &self,
     app_handle: &AppHandle<R>,
     on_menu_event: Option<GlobalMenuEventListener<AppHandle<R>>>,
-    on_tray_event: Option<GlobalTrayIconEventListener<TrayIcon<R>>>,
+    on_tray_icon_event: Option<GlobalTrayIconEventListener<TrayIcon<R>>>,
   ) {
     if let Some(handler) = on_menu_event {
       app_handle
@@ -207,7 +301,7 @@ impl<R: Runtime> TrayIcon<R> {
         .push(handler);
     }
 
-    if let Some(handler) = on_tray_event {
+    if let Some(handler) = on_tray_icon_event {
       app_handle
         .manager
         .inner
@@ -247,7 +341,10 @@ impl<R: Runtime> TrayIcon<R> {
   }
 
   /// Register a handler for this tray icon events.
-  pub fn on_tray_event<F: Fn(&TrayIcon<R>, TrayIconEvent) + Sync + Send + 'static>(&self, f: F) {
+  pub fn on_tray_icon_event<F: Fn(&TrayIcon<R>, TrayIconEvent) + Sync + Send + 'static>(
+    &self,
+    f: F,
+  ) {
     self
       .app_handle
       .manager